Add two themes with their assets to ThemesFrame
